formatting

+++ src/flowMC/strategy/take_steps.py
@@ -6,7 +6,6 @@ from flowMC.resource.logPDF import LogPDF
 from flowMC.strategy.base import Strategy
 from jaxtyping import Array, Float, PRNGKeyArray
 import jax
-import jax.numpy as jnp
 import equinox as eqx
 from abc import abstractmethod
 
@@ -107,7 +106,9 @@ class TakeSteps(Strategy):
 
         positions = positions[:, :: self.thinning]
         log_probs = log_probs[:, :: self.thinning]
-        do_accepts = do_accepts[:, :: self.thinning].astype(acceptance_buffer.data.dtype)
+        do_accepts = do_accepts[:, :: self.thinning].astype(
+            acceptance_buffer.data.dtype
+        )
 
         position_buffer.update_buffer(positions, self.current_position)
         log_prob_buffer.update_buffer(log_probs, self.current_position)
